Factor zephyr helpers out of zephyr_whatsnew.py
Factor the Zephyr repo analyzer and friends into their own module,
zephyr_helpers.py. Move the test cases into their own file as well.

The new way to run tests is:

py.test test_*.py

The behavior of the script is unchanged, but the Foundries.io specific
bits remained in zephyr_whatsnew.py. The zephyr_helpers.py module is
agnostic to the organization maintaining the downstream.
