+++ test_zephyr_helpers.py
@@ -1,0 +1,167 @@
+from zephyr_helpers import shortlog_area
+
+
+def test_shortlog_to_area():
+    # Some areas, and shortlogs that should map to them
+    area_shortlog_expected = [
+        # Cases where we expect to match a shortlog to a particular area.
+
+        ('Arches', 'ARM: stm32f0: fix syscfg mapping to fix EXTI config'),
+        ('Arches', 'x86: mmu: kernel: Validate existing APIs'),
+        ('Arches', 'arch: x86: fix jailhouse build'),
+        ('Arches', 'arm: implement API to validate user buffer'),
+        ('Arches',
+         'xtensa/asm2: Add a _new_thread implementation for asm2/switch'),
+        ('Arches', 'esp32: Set CPU pointer on app cpu at startup'),
+        ('Arches', 'native_posix: Be more precise with stop-at'),
+        ('Arches', 'riscv: Add device tree support to pulpino'),
+        ('Bluetooth', 'Bluetooth: Mesh: Fix typo in Kconfig help message'),
+        ('Bluetooth',
+         'bt: hci driver over spi: BlueNRG-MS read until IRQ pin goes low'),
+        ('Boards', 'boards: nios2: altera_max10: cleanup board documentation'),
+        ('Boards', 'mimxrt1050_evk: Configure an lpspi instance and pins'),
+        ('Build', 'cmake: Fix the dependency between qemu and the elf file'),
+        ('Build', 'kconfig: 802154: nrf: Fix kconfig'),
+        ('Build', 'gen_syscall_header: create dummy handler refs'),
+        ('Build', 'Revert "cmake: add zephyr_cc_option_nocheck"'),
+        ('Build', 'gen_isr_tables: Minor refactoring'),
+        ('Build', 'toolchain: organise toolchain/compiler files'),
+        ('Build', 'clang/llvm: add initial configuration file for clang'),
+        ('Build', 'genrest: Mention implicit default values'),
+        ('Build',
+         'isr_tables: Simplify how the spurious irq function address is found'),  # noqa: E501
+        ('Build', 'menuconfig: Fix searching for nonexistent objects'),
+        ('Build', 'toolchains: add xtools support for ARC'),
+        ('Continuous Integration', 'sanitycheck: Flush stdout in info()'),
+        ('Continuous Integration', 'ci: verify author identity'),
+        ('Continuous Integration',
+         'coverage: build with -O0 to get more information'),
+        ('Continuous Integration',
+         'gitlint: do not allow title-only commit messages'),
+        ('Cryptography', 'crypto: Update TinyCrypt to 0.2.8'),
+        ('Cryptography',
+         'crypto: config: config-coap: ' +
+         'add CONFIG for setting max content length'),
+        ('Cryptography',
+         'mbedtls: Kconfig: Re-organize to enable choosing an mbedtls impl.'),
+        ('Device Tree',
+         'include: dt-bindings: stm32_pinctrl: Add ports I, J, K'),
+        ('Device Tree', 'dt: nrf52840: remove 0x from USBD address'),
+        ('Device Tree', 'dts/arm: Move i2c2 node inside stm32fxxx dtsi file'),
+        ('Device Tree', 'dts/arm/st: fix dts inclusion for stm32f334'),
+        ('Drivers', 'drivers: serial: stm32: report only unmasked irq'),
+        ('Drivers', 'flash: stm32l4x: fix build'),
+        ('Drivers', 'gpio: Introduce mcux igpio shim driver'),
+        ('Drivers', 'clock_control: Introduce mcux ccm driver'),
+        ('Drivers', 'serial: Add another instance to the mcux lpuart driver'),
+        ('Drivers', 'drivers/ieee802154_kw41z: Fix interrupt priority'),
+        ('Drivers',
+         'usb: netusb: Use lower addresses for default endpoint config'),
+        ('Drivers', 'device: cleanup header layout'),
+        ('Drivers', 'uart: fixing pin range being too tight for the nrf52840'),
+        ('Drivers',
+         'device.h: doc: Refactor to keep documentation infront of impl.'),
+        ('Drivers', 'sensors/lsm5dsl: Fix SPI API usage'),
+        ('Drivers',
+         "hid: core: truncated wLength if it doesn't match report descriptor "
+         "size"),
+        ('Drivers',
+         'uart_pipe: re-work the RX function to match the API '
+         'and work with USB.'),
+        ('Drivers', 'netusb: rndis: Add more debugs'),
+        ('Drivers', 'can: Add can support for STM32L432'),
+        ('Drivers',
+         'subsys: usb/class/hid: make interrupt endpoint size configurable'),
+        ('Documentation', 'doc/dts: Update to reflect new path locations'),
+        ('Documentation', 'doc: boards: v2m_beetle: fix conversion to cmake'),
+        ('Documentation', 'doxygen: ignore misc/util.h'),
+        ('External', 'ext: hal: altera: Add Altera HAL README file'),
+        ('External', 'ext/hal: stm32cube: Update STM32F0 README file'),
+        ('Firmware Update',
+         'dfu: replace FLASH_ALIGN with FLASH_WRITE_BLOCK_SIZE'),
+        ('Firmware Update', 'subsys: mgmt: SMP protocol for mcumgr.'),
+        ('Storage', 'disk: delete the GET_DISK_SIZE IOCTL.'),
+        ('Storage',
+         'subsys: fcb: Check for mutex lock failure when walking FCB'),
+        ('Kernel', 'kernel: stack: add -fstack-protector-all without checks'),
+        ('Kernel',
+         'Revert "kernel: stack: add -fstack-protector-all without checks"'),
+        ('Kernel', 'poll: k_poll: Document -EINTR return'),
+        ('Kernel', 'mempool: add assertion for calloc bounds overflow'),
+        ('Kernel',
+         'syscalls: REVERTME: clean up warnings when building unit tests'),
+        ('Kernel',
+         ('work_q: Correctly clear pending flag in delayed work queue, '
+          'update docs')),
+        ('Kernel', 'init.h: Fix english in comment'),
+        ('Kernel', 'userspace: automatic resource release framework'),
+        ('Kernel', 'k_queue: allow user mode access via allocators'),
+        ('Kernel', 'k_poll: expose to user mode'),
+        ('Kernel', 'subsys: app_memory: Fixed incorrect linker syntax.'),
+        ('Libraries',
+         'libc: some architectures do not require baremetal libc'),
+        ('Libraries', 'lib: move ring_buffer from misc/ to lib/'),
+        ('Libraries', 'ring_buffer: remove broken object_tracing support'),
+        ('Libraries', 'lib/rbtree: Fix crash condition with empty trees and rb_min/max()'),  # noqa: E501
+        ('Libraries', 'lib/posix: Port wait_q usage to new API'),
+        ('Logging', 'logging: Add internal thread for log processing'),
+        ('Logging', 'logger: Rename SHOW_COLOR and FORMAT_TIMESTAMP options'),
+        ('Maintainers', 'CODEOWNERS: misc updates'),
+        ('Maintainers', 'CODEOWNERS.rst: misc updates'),
+        ('Miscellaneous',
+         'printk: Add padding support to string format specifiers'),
+        ('Miscellaneous',
+         'version: fix version handling without extra_version set'),
+        ('Miscellaneous', 'misc: Use braces in infinite for loop'),
+        ('Networking', 'net: if: fix ND reachable calculation'),
+        ('Networking', 'net/ieee802154: Make RAW mode generic'),
+        ('Networking', 'openthread: Use ccache when enabled'),
+        ('Networking', 'slip: fix a bug when in non-TAP mode.'),
+        ('Power Management',
+         'subsys: power: Add OS managed Power Management framework'),
+        ('Samples', 'samples: echo_server: Test the nrf build in CI'),
+        ('Samples',
+         'samples/xtensa-asm2: Unit test for new Xtensa assembly primitives'),
+        ('Scripts',
+         'scripts: runner: nrfjprog: remove BOARD environment requirement'),
+        ('Scripts', "scripts: jlink: Don't reset after load"),
+        ('Scripts', 'runner: nrfjprog: Improve error messages'),
+        ('Scripts',
+         'scripts/dts: '
+         'Use 4-spaces tabs instead of 2-space tabs in devicetree.py'),
+        ('Scripts',
+         'script/dts: Remove unnecessary empty return on functions'),
+        ('Scripts', 'gen_syscalls.py: fix include issue'),
+        ('Scripts', 'gen_syscall_header.py: fix include issue'),
+        ('Scripts', 'kconfiglib: Update to 2259d353426f1'),
+        ('Scripts', 'Coccinelle: Add support for Coccinelle infrastructure'),
+        ('Scripts', 'west: runner: fix naming and paths to xt-ocd'),
+        ('Storage', 'subsys: settings: fix fcb back-end initialization'),
+        ('Storage', 'subsys: fs/nvs: improve syslog messages'),
+        ('Testing', 'tests: use cmake to build object benchmarks'),
+        ('Testing',
+         'tests: mem_pool: ' +
+         'Fixed memory pool test case failure on quark d2000.'),
+        ('Testing',
+         'tests/kernel/mem_protect/userspace: ' +
+         'test that _k_neg_eagain is in rodata'),
+        ('Testing', 'unittest: Support EXTRA_*_FLAGS'),
+        ('Testing', 'testing: add option to generate coverage reports'),
+        ('Testing',
+         'tracing: don\'t include kernel_structs.h from tracing_sysview.h'),
+
+        # Cases we explicitly do not want to match, and why:
+
+        # Tree-wide change with no particular area.
+        (None, 'Introduce cmake-based rewrite of KBuild'),
+
+        # Should have been 'boards: mimxrt1050_evk' or so.
+        (None, 'mimxrt1050_evk'),
+
+        # Should have been 'arm: _setup_new_thread' or so.
+        (None, '_setup_new_thread: fix crash on ARM'),
+    ]
+
+    for expected, shortlog in area_shortlog_expected:
+        actual = shortlog_area(shortlog)
+        assert actual == expected, 'failed on: "{}"'.format(shortlog)
